requirements.txt
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
pypdf
faiss-cpu
google-search-results
python-multipart
tiktoken
openai
pydantic